More macro renaming to match Bevy
- `add_resource` -> `insert_resource`
- `register_event` -> `add_event`
