+++ Cargo.toml
@@ -12,3 +12,5 @@ exclude = [".direnv/", ".vscode/", "Cargo.lock", "*.nix", "flake.lock", ".envrc"
 bevy_app = { version = "0.16.0-rc.2", default-features = false }
 bevy_ecs = { version = "0.16.0-rc.2", default-features = false }
 bevy_log = { version = "0.16.0-rc.2", default-features = false }
+bevy_state = { version = "0.16.0-rc.2", default-features = false, features = ["bevy_app"] }
+bevy = { version = "0.16.0-rc.2", default-features = false, features = ["bevy_log"] }
