+++ bevy-butler-proc-macro/src/insert_resource/structs.rs
