+++ bevy-butler/tests/insert_resource/generic_resource.rs
@@ -7,24 +7,24 @@ use wasm_bindgen_test::wasm_bindgen_test;
 use crate::common::log_plugin;
 
 #[derive(Resource, Default)]
-#[add_resource(plugin = MyPlugin)]
-struct Marker(bool);
+#[insert_resource(plugin = MyPlugin, generics = <bool>)]
+struct Marker<T>(T);
 
 #[derive(Resource)]
-#[add_resource(plugin = MyPlugin, init = Message("Hello, world!".to_string()))]
-struct Message(String);
+#[insert_resource(plugin = MyPlugin, init = Message("Hello, world!".to_string()))]
+struct Message<T>(T);
 
 #[butler_plugin]
 struct MyPlugin;
 
 #[add_system(plugin = MyPlugin, schedule = Startup)]
-fn get_and_print_message(message: Res<Message>, mut marker: ResMut<Marker>) {
+fn get_and_print_message(message: Res<Message<String>>, mut marker: ResMut<Marker<bool>>) {
     info!("Resource message: {}", message.0);
     marker.0 = true;
 }
 
 #[add_system(plugin = MyPlugin, schedule = PostStartup)]
-fn assert_marker(marker: Res<Marker>) {
+fn assert_marker(marker: Res<Marker<bool>>) {
     assert!(marker.0);
 }
